requirements.txt
annotated-doc==0.0.4
anyio==4.12.1
asttokens==3.0.1
Cartopy==0.25.0
certifi==2026.2.25
click==8.3.1
contourpy==1.3.2
cuda-bindings==12.9.4
cuda-pathfinder==1.4.0
cycler==0.12.1
decorator==5.2.1
exceptiongroup==1.3.1
executing==2.2.1
filelock==3.25.0
fonttools==4.61.1
fsspec==2026.2.0
geoclip==1.2.0
geographiclib==2.1
geopy==2.4.1
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.11
ipython==8.38.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
packaging @ file:///home/task_176104877067765/conda-bld/packaging_1761049113113/work
pandas==2.3.3
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
POT==0.9.6.post1
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.3.2
pyproj==3.7.1
pyshp==3.0.3
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.2.28
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
traitlets==5.14.3
transformers==5.3.0
triton==3.6.0
typer==0.24.1
typing_extensions==4.15.0
tzdata==2025.3
wcwidth==0.6.0
